requirements.txt
openai-whisper>=20231117
moviepy>=1.0.3
srt>=3.5.3
gradio>=4.19.2
yt-dlp>=2024.3.10
ffmpeg-python>=0.2.0 